Move ThreadDumps and StackTraceUtils to their own package
Also, wrap the calls with isTraceEnabled() for performance reasons.
