+++ atlasdb-commons/src/main/java/com/palantir/util/debug/ThreadDumps.java
@@ -13,8 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.palantir.util;
-
+package com.palantir.util.debug;
 
 import java.io.IOException;
 import java.lang.management.ManagementFactory;
