+++ atlasdb-config/src/main/java/com/palantir/atlasdb/factory/ServiceDiscoveringAtlasSupplier.java
@@ -30,7 +30,7 @@ import com.palantir.atlasdb.keyvalue.api.KeyValueService;
 import com.palantir.atlasdb.spi.AtlasDbFactory;
 import com.palantir.atlasdb.spi.KeyValueServiceConfig;
 import com.palantir.timestamp.TimestampService;
-import com.palantir.util.ThreadDumps;
+import com.palantir.util.debug.ThreadDumps;
 
 public class ServiceDiscoveringAtlasSupplier {
     private static final Logger log = LoggerFactory.getLogger(ServiceDiscoveringAtlasSupplier.class);
@@ -58,9 +58,11 @@ public class ServiceDiscoveringAtlasSupplier {
     }
 
     public TimestampService getTimestampService() {
-        log.trace("Fetching timestamp service from thread {}. This should only happen once.",
-                Thread.currentThread().getName());
-        log.trace("Thread dump: " + ThreadDumps.programmaticThreadDump());
+        if (log.isTraceEnabled()) {
+            log.trace("Fetching timestamp service from thread {}. This should only happen once.",
+                    Thread.currentThread().getName());
+            log.trace("Thread dump: " + ThreadDumps.programmaticThreadDump());
+        }
 
         return timestampService.get();
     }
